python/jaeger/database-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-exporter-otlp==1.22.0